requirements.txt
fastapi==0.121.0
numpy==2.3.4
pandas==2.3.3
Pillow==12.0.0
scikit-learn==1.7.2
uvicorn==0.38.0
